+++ SA.HM.Presentation/IIM.Presentation.Website/Common/HMReport.Master.cs
@@ -648,7 +648,7 @@ namespace HotelManagement.Presentation.Website.Common
         }
         public void MessageCount()
         {
-            Int16 TotalUnreadMessage = 0;
+            Int64 TotalUnreadMessage = 0;
 
             CommonMessageDA messageDa = new CommonMessageDA();
             List<CommonMessageDetailsBO> messageDetails = new List<CommonMessageDetailsBO>();
